refactor: feature 폴더명을 camelCase로 정리 (#125)
